fix(lineup-ingest): 打通全管线，修复多环节卡点
- types.ts: startSec 去掉 .int()，支持小数秒
- check-sources.ts: 递归遍历 sources/ 子目录
- video-vlm.ts: callGateway 改用 video_url 类型，优化错误信息，兼容硅基流动 Qwen3-VL
- review: map/agent/side/site 改为下拉选项，切换时自动更新 id；/api/config 加入地图数据
- promote.ts: 修复帧路径 join 错误；自动注册 data/lineups 新文件到 ALL_LINEUPS
- ingest-images.ts: 加载 .env.local 读取 CDN_BASE
- README: 更新完整管线 7 步流程 + 环境变量表
- 新增: sources/ascent/cypher/, sources/summit/jett1/, data/lineups/ascent-cypher.json, summit-jett.json

+++ packages/lineup-ingest/src/types.ts
@@ -6,7 +6,7 @@ import type { LlmExtractor } from './extractors/types';
 
 /** 手抄自视频简介的时间轴一行：起点 + 小标题（结束点由下一段推） */
 export const manualSegmentSchema = z.object({
-  startSec: z.number().int().nonnegative(),
+  startSec: z.number().nonnegative(),
   title: z.string().min(1),
 });
 export type ManualSegment = z.infer<typeof manualSegmentSchema>;
